# entertainmentEmporiumBackend/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Entertainment Emporium Backend API 
  description: A movies and tv shows rating API 
  contact:
    name: Francesca 
    email: passmoref@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://hexagonprofile-maximumalex-3000.codio-box.uk/api/v1'
    description: Development server
...
paths:
...
  /tv/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the tv show
    get:
      tags:
        - Tv
      description:
        Get tv show by id.
      responses:
        '200':
          description: View of a tv shows record.
        '404':
          description: Tv show was not found in the DB.
      
    put:
      tags:
        - Tv
      description:
        Update tv show by ID.
      responses:
        '200':
          description: Update confirmation and link to record.
        '404':
          description: Tv show was not found in the DB.
        '401':
          description: Admins can only update tv shows
        '403':
          description: Requester forbidden to update this resource

    delete:
      tags:
        - Tv
      description:
        Delete tv show by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only admin users can delete tv show.
        '404':
          description: Tv show was not found in the DB.
        '403':
          description: Requester forbidden to delete this resource
  /tv:
    summary: Represents an individual tv show
    description: >
      This resource represents an individual tv show in the system.
      Each tv show is identified by a numeric `id`.
    post:
      tags:
        - Tv
      description:
        Create a new tv show.
      requestBody:
        description: All new tv show data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./tv.json#/definitions/tv
      responses:
        '201':
          description: Tv show created
          content:
            application/json:
              schema:
                $ref: ./tv.json#/definitions/tvCreated
...
  /tv/{title}:
    parameters:
      - in: path
        name: title
        schema:
          type: string
        required: true
        description: Title of the tv show 
    get:
      tags:
        - Tv
      description:
        Get tv show by the title.
      responses:
        '200':
          description: View of the tv show record.
        '404':
          description: Tv show was not found in the DB.
...
  /tvReviews/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the tv review
    get:
      tags:
        - Tv Reviews
      description:
        Get tv review by id.
      responses:
        '200':
          description: View of a tv reviews record 
        '404':
          description: Tv review was not found in the DB.
    put:
      tags:
        - Tv Reviews
      description:
        Update tv reviews by ID.
      responses:
        '200':
          description: Update confirmation.
        '404':
          description: Tv review was not found in the DB.
        '401':
          description: Users can only update their own tv reviews
        '403':
          description: Requester forbidden to update this resource

    delete:
      tags:
        - Tv Reviews
      description:
        Delete tv reviews by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only users can delete their own tv reviews.
        '404':
          description: Tv review was not found in the DB.
        '403':
          description: Requester forbidden to delete this resource
  /tvReviews:
    summary: Represents an individual tv review
    description: >
      This resource represents an individual tv review in the system.
      Each tv review is identified by a numeric `id`.
    post:
      tags:
        - Tv Reviews
      description:
        Create a new tv review account.
      requestBody:
        description: All new tv review data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./tvReviews.json#/definitions/tvReview
      responses:
        '200':
          description: Tv created
          content:
            application/json:
              schema:
                $ref: ./tvReviews.json#/definitions/reviewCreated